Moved extensions into separate files
